gu33: updated grid format

+++ course/format/grid/classes/observer.php
@@ -20,8 +20,10 @@
  * @package    course/format
  * @subpackage grid
  * @category   event
+ * @version    See the value of '$plugin->version' in version.php.
  * @copyright  &copy; 2017-onwards G J Barnard based upon work done by Marina Glancy.
- * @author     G J Barnard - gjbarnard at gmail dot com, about.me/gjbarnard and {@link http://moodle.org/user/profile.php?id=442195}
+ * @author     G J Barnard - {@link http://about.me/gjbarnard} and
+ *                           {@link http://moodle.org/user/profile.php?id=442195}
  * @author     Based on code originally written by Paul Krix and Julian Ridden.
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
@@ -41,15 +43,18 @@ class format_grid_observer {
      * @param \core\event\course_content_deleted $event
      */
     public static function course_content_deleted(\core\event\course_content_deleted $event) {
-        global $DB;
-        /* Delete any images associated with the course.
-           Done this way so will work if the course has
-           been a grid format course in the past even if
-           it is not now. */
-        $courseformat = format_grid::get_instance($event->objectid);
-        $courseformat->delete_images();
-        unset($courseformat);  // Destruct.
+        if (class_exists('format_grid', false)) {
+            // If class format_grid was never loaded, this is definitely not a course in 'Grid' format.
+            global $DB;
+            /* Delete any images associated with the course.
+               Done this way so will work if the course has
+               been a grid format course in the past even if
+               it is not now. */
+            $courseformat = format_grid::get_instance($event->objectid);
+            $courseformat->delete_images();
+            unset($courseformat);  // Destruct.
 
-        $DB->delete_records("format_grid_summary", array("courseid" => $event->objectid));
+            $DB->delete_records("format_grid_summary", array("courseid" => $event->objectid));
+        }
     }
 }
